fix: determine user's home directory on remote robustly (#49)
* fix(executor): handle scenario when `conn_opts` is not passed

* feat(provider): determine home directory robustly

We used to hardcode `~` to denote the home directory and hoped that would get resolved on the remote host. This lead to issues during docker support as `~` could not be resolved correctly in all cases (or maybe I did not know how to do it correctly with the correct shell escaping). But nonetheless, we have now relaxed this consideration and now determine the remote home dynamically during each new host's first setup run. This is done by running `echo $HOME` against the remote host and parsing the response to get the path. This path is saved as part of the configuration (same as earlier) and re-used during later runs.

* chore(build): auto-generate vimdoc

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>