feat: improve manual ssh input and remote connection check
Earlier, we created a workspace configuration without checking if we are able to connect to the remote or not. Now, we first check if we are able to connect or not. If we are able to connect, only then we create a new configuration. Additionally, we improved the auto host detection logic of ssh manual input. If only a single input is provided, assume that it is the host URI.