fix(scripts): Explicitly specify bash command to run scripts (#156)
Nix-based systems mess up the interpreter path to make it useless when
it gets copied over to the remote system. So, instead now, we manually
specify the shell to execute it under. So far, I cannot think of any
drawbacks of using this except the need to be explicit which is honestly
not a bad thing. We'll see how this pans out though.